requirements.txt
pyyaml==6.0.2
jsonschema==4.22.0
orjson==3.10.7
rich==13.7.1
